requirements.txt
Django==4.2 #Framework
django-cors-headers==4.2.0 # Habilita el soporte para Cross-Origin Resource Sharing (CORS) en aplicaciones Django.
django-environ==0.10.0 # Simplifica la gestión de la configuración mediante variables de entorno.
djangorestframework==3.14.0 #Apis rest
transformers==4.39.3
torch>=2.1.0